# lsv-backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/lesson/{id}": {
      "get": {
        "operationId": "UsersController_getLessonById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/quiz/{id}": {
      "put": {
        "operationId": "QuizController_updateQuiz",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      },
      "delete": {
        "operationId": "QuizController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      }
    },
    "/quiz/{quizId}": {
      "get": {
        "operationId": "QuizController_getQuizById",
        "parameters": [
          {
            "name": "quizId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Quiz"
        ]
      }
    },
...
    "/quiz-variants/lesson-variant/{lessonVariantId}": {
      "get": {
        "operationId": "QuizVariantController_getQuizVariants",
        "parameters": [
          {
            "name": "lessonVariantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuizVariant"
        ]
      }
    },
    "/quiz-variants": {
      "post": {
        "operationId": "QuizVariantController_createQuizVariant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizVariantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "QuizVariant"
        ]
      }
    },
    "/quiz-variants/{id}": {
      "delete": {
        "operationId": "QuizVariantController_deleteQuizVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "QuizVariant"
        ]
      }
    },
    "/lesson/by-language/{languageId}": {
      "get": {
        "operationId": "LessonController_getLessonsByLanguageAdmin",
        "parameters": [
          {
            "name": "languageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
    "/lesson/by-language/{languageId}/with-submissions": {
      "get": {
        "operationId": "LessonController_getLessonsByLanguageWithSubmissions",
        "parameters": [
          {
            "name": "languageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
    "/lesson": {
      "get": {
        "operationId": "LessonController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      },
      "post": {
        "operationId": "LessonController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
    "/lesson/{id}/image": {
      "post": {
        "operationId": "LessonController_uploadImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
    "/lesson/{id}/with-quizzes": {
      "get": {
        "operationId": "LessonController_getLessonWithQuizzes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
    "/lesson/{id}/quizzes": {
      "get": {
        "operationId": "LessonController_getQuizzesByLessonId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
    "/lesson/{id}": {
      "put": {
        "operationId": "LessonController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      },
      "delete": {
        "operationId": "LessonController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      },
      "get": {
        "operationId": "LessonController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
    "/lesson/{id}/variants": {
      "get": {
        "operationId": "LessonController_getLessonVariants",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      },
      "post": {
        "operationId": "LessonController_createLessonVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonVariantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
    "/lesson/{id}/variants/{variantId}": {
      "get": {
        "operationId": "LessonController_getLessonVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      },
      "put": {
        "operationId": "LessonController_updateLessonVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonVariantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      },
      "delete": {
        "operationId": "LessonController_deleteLessonVariant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
    "/lesson/regional/{id}": {
      "get": {
        "operationId": "LessonController_getRegionalLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "regionId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lesson"
        ]
      }
    },
...
  },
  "info": {
    "title": "API LSV",
    "description": "Esquema OpenAPI de la API LSV",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "QuizDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateQuizVariantDto": {
        "type": "object",
        "properties": {}
      },
      "CreateLessonDto": {
        "type": "object",
        "properties": {}
      },
      "CreateLessonVariantDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}